fix: 별점 정보 Integer -> Float 수정

+++ src/main/java/dku/capstone/foodlog/repository/PlaceRepositoryCustomImpl.java
@@ -39,7 +39,7 @@ public class PlaceRepositoryCustomImpl implements PlaceRepositoryCustom{
         return !categoryList.isEmpty()? place.category.in(categoryList) : null;
     }
 
-    private BooleanExpression ratingGoe(Integer rating) {
+    private BooleanExpression ratingGoe(Float rating) {
         return rating != null? placePost.averageRating.goe(rating) : null;
     }
